+++ frontend/src/scenes/notebooks/Nodes/components/NotebookNodeTitle.tsx
@@ -1,6 +1,6 @@
 import { useActions, useValues } from 'kea'
 import posthog from 'posthog-js'
-import { KeyboardEvent, useEffect, useMemo, useState } from 'react'
+import { KeyboardEvent, useEffect, useState } from 'react'
 
 import { LemonInput, LemonTag, Tooltip } from '@posthog/lemon-ui'
 
@@ -11,60 +11,11 @@ import { isHogQLQuery } from '~/queries/utils'
 import { NotebookNodeType } from '../../types'
 import { notebookNodeLogic } from '../notebookNodeLogic'
 
-const collectNodeIndices = (
-    content: Record<string, any> | null | undefined,
-    predicate: (node: Record<string, any>) => boolean
-): Map<string, number> => {
-    if (!content || typeof content !== 'object') {
-        return new Map()
-    }
-
-    const nodeIndices = new Map<string, number>()
-    let currentIndex = 0
-
-    const walk = (node: Record<string, any> | null | undefined): void => {
-        if (!node || typeof node !== 'object') {
-            return
-        }
-
-        if (predicate(node)) {
-            const nodeId = node.attrs?.nodeId
-            if (nodeId) {
-                currentIndex += 1
-                nodeIndices.set(nodeId, currentIndex)
-            }
-        }
-
-        if (Array.isArray(node.content)) {
-            node.content.forEach(walk)
-        }
-    }
-
-    walk(content)
-    return nodeIndices
-}
-
 export function NotebookNodeTitle(): JSX.Element {
-    const { isEditable, content } = useValues(notebookLogic)
+    const { isEditable, pythonNodeIndices, sqlNodeIndices } = useValues(notebookLogic)
     const { nodeAttributes, title, titlePlaceholder, isEditingTitle, nodeType } = useValues(notebookNodeLogic)
     const { updateAttributes, toggleEditingTitle } = useActions(notebookNodeLogic)
     const [newValue, setNewValue] = useState('')
-
-    const pythonNodeIndices = useMemo(
-        () => collectNodeIndices(content, (node) => node.type === NotebookNodeType.Python),
-        [content]
-    )
-    const sqlNodeIndices = useMemo(
-        () =>
-            collectNodeIndices(
-                content,
-                (node) =>
-                    node.type === NotebookNodeType.Query &&
-                    (isHogQLQuery(node.attrs?.query) ||
-                        (node.attrs?.query.source && isHogQLQuery(node.attrs.query.source)))
-            ),
-        [content]
-    )
 
     const isPythonNode = nodeType === NotebookNodeType.Python
     const isSqlNode =
